pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "playlist-transfer"
version = "0.1.0"
description = "Transfer playlists between music streaming platforms"
requires-python = ">=3.7"
dependencies = [
    "spotipy>=2.23.0",
    "tidalapi>=0.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
playlist-transfer = "playlist_transfer.main:main"

[tool.setuptools]
packages = ["playlist_transfer", "playlist_transfer.services"]
